+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- イベント予約集計時にevServiceStatus,onlineSalesStatus,tourNumberを更新しないように調整
+
 ### Deprecated
 
 ### Removed